# docs/swagger.yaml
definitions:
  app.ResponseModel:
    properties:
      data:
        type: object
      message:
        description: Code int         `json:"code"`
        type: string
    type: object
...
  models.AddUserForm:
    properties:
      client_id:
        type: string
      company_id:
        type: integer
      created_by:
        type: string
      data_permission:
        type: string
      email_addr:
        type: string
      file_id:
        type: string
      handphone_no:
        type: string
      level_no:
        type: integer
      name:
        type: string
      role_id:
        type: string
      user_name:
        type: string
      user_status:
        type: integer
    type: object
...
  models.EditUserForm:
    properties:
      client_id:
        type: string
      company_id:
        type: integer
      data_permission:
        type: string
      email_addr:
        type: string
      file_id:
        type: string
      handphone_no:
        type: string
      level_no:
        type: integer
      name:
        type: string
      role_id:
        type: string
      updated_by:
        type: string
      user_name:
        type: string
      user_status:
        type: integer
    type: object
...
  models.ResponseModelList:
    properties:
      all_column:
        type: string
      data:
        type: object
      define_column:
        type: string
      define_size:
        type: string
      last_page:
        type: integer
      message:
        type: string
      page:
        type: integer
      total:
        type: integer
    type: object
...
info:
  contact:
    email: nuryantofattih@gmail.com
    name: Nuryanto
    url: https://www.linkedin.com/in/nuryanto-1b2721156/
  description: Backend REST API for golang starter
  license: {}
  title: Starter
  version: "1.0"
paths:
...
  /api/user:
    delete:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.ResponseModel'
      security:
      - ApiKeyAuth: []
      summary: GetList SaUser
      tags:
      - User
    get:
      parameters:
      - description: Page
        in: query
        name: page
        required: true
        type: integer
      - description: PerPage
        in: query
        name: perpage
        required: true
        type: integer
      - description: Search
        in: query
        name: search
        type: string
      - description: InitSearch
        in: query
        name: initsearch
        type: string
      - description: SortField
        in: query
        name: sortfield
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseModelList'
      security:
      - ApiKeyAuth: []
      summary: GetList SaUser
      tags:
      - User
    post:
      parameters:
      - description: 'req param #changes are possible to adjust the form of the registration
          form from frontend'
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/models.AddUserForm'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.ResponseModel'
      security:
      - ApiKeyAuth: []
      summary: Add User
      tags:
      - User
  /api/user/{id}:
    get:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.ResponseModel'
      security:
      - ApiKeyAuth: []
      summary: GetById SaUser
      tags:
      - User
    put:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: 'req param #changes are possible to adjust the form of the registration
          form from frontend'
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/models.EditUserForm'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.ResponseModel'
      security:
      - ApiKeyAuth: []
      summary: Update User
      tags:
      - User
  /api/user/permission:
    get:
      parameters:
      - description: ClientID
        in: query
        name: client_id
        required: true
        type: string
      - description: UserID
        in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.ResponseModel'
      security:
      - ApiKeyAuth: []
      summary: Get string Array Permission User
      tags:
      - User
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"